Update heatmap title display logic in GraphView.vue
Changed the condition to display the heatmap title in GraphView.vue. It now only displays when there is more than one heatmap, instead of displaying when there is at least one. This small change will improve the handling and presentation of heatmap data in cases where only a single heatmap is present.